requirements.txt
# Core Python Libraries
flask==2.1.0
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
tensorflow==2.7.0

# Google Cloud Libraries
google-cloud-core==2.2.2
google-cloud-service-management==1.3.0
google-auth==2.3.3
google-cloud-storage==2.1.0
google-cloud-bigquery==2.31.0
google-cloud-pubsub==2.10.0
google-cloud-dataflow==0.17.0
google-cloud-aiplatform==1.16.0

# Additional Diagnostic Libraries
scapy==2.4.5
psutil==5.8.0
requests==2.26.0